Fix breakage from removal of Arduino_TensorFlowLite dependency from Library Manager
The "Arduino_TensorFlowLite" library is a dependency of this project.

This dependency was previously included in the Arduino Library Manager index, which was used by the project in the
following ways:

- Specified as dependency in metadata to provide automated installation along with TinyTrainable
- Referenced by name alone in the library dependencies configuration of the "Arduino Compile" GitHub Actions workflow
- Link to the generated library reference on arduino.cc

The library was removed from the Arduino Library Manager index at the request of the TensorFlow Lite Micro maintainers,
which broke all three of the usages listed above.

That breakage is hereby repaired:

All libraries listed in the `depends` field of the library.properties metadata file must be available for installation
from Library Manager. The presence of Arduino_TensorFlowLite on that list causes the library installation to fail with
Arduino IDE 2.x, or `arduino-cli lib install TinyTrainable`. The library is removed from the list.

The configuration of the "Arduino Compile" GitHub Actions workflow is adjusted to install the library by cloning the
official tensorflow/tflite-micro-arduino-examples library. Since that repository does not currently provide any tags,
the workflow is configured to use the version of the library from the tip of the main branch.

The readme is updated to specify that the Arduino_TensorFlowLite library must be installed manually, and the link
updated from the obsolete generated reference page to the official repository for the library, which contains manual
installation instructions.